Rename: SkillSelector Component를 common 폴더로 이동
