+++ src/components/organism/CreateStudy/FormElements/index.tsx
@@ -4,7 +4,7 @@ import { ChangeEvent, MouseEvent } from 'react';
 import { Label } from '@components/atoms/label/Label';
 import { Input } from '@components/ui/Input/Input';
 import { Select } from '@components/ui/Select/Select';
-import { SkillSelector } from '@components/domain/SkillSelector';
+import { SkillSelector } from '@components/common/selector/SkillSelector';
 import { Checkbox } from '@components/ui/Checkbox/Checkbox';
 import VerticalLine from '@assets/svg/recentLine.svg';
 import DownArrow from '@assets/svg/recentIcon.svg';
